Remove 'tense' from the reference to 'Classes' section
The current text refers to the classes section in the 'past tense' when in the handbook flow it is in the future. I am proposing that the sentence be changed to 'tense neutral' to account for the reading sequence changes, if any, that happen, or if the reader follows a different flow.