Link to Tuple Types section of Object Types page
A search for 'tuple' on the website has the deprecated handbook basic types page as the only documentation result. That deprecated page links back to this one so there was is no way to find the current Tuple Types documentation in Object Types from the website.

Unfortunately a search for 'tuple types' shows no documentation results. Perhaps documentation sections (rather than just pages) could be indexed in Algolia.