firstElement examples should take into consideration undefined - revised
The behavior of the return type of 'firstElement' will differ based on the setting of https://www.typescriptlang.org/tsconfig#noUncheckedIndexedAccess. When this is set to 'true' the original example will not compile. By being more explicit in the return type (i.e. using:  T | undefined) the code compiles regardless of that setting and beyond that expresses intent (even) more clearly.