Updated false information in Decorators.md
Below this example decorator function, the documentation states: "When @sealed is executed, it will seal both the constructor and its prototype which would not allow the class to be sub-classed at runtime."

```js
function sealed(constructor: Function) {
  Object.seal(constructor);
  Object.seal(constructor.prototype);
}
```
However, this information is incorrect as far as I'm concerned. This code doesn't prevent classes from subclassing ones decorated with `@sealed`. It only prevents any further modification from being made to `BugReport.prototype` and to the BugReport constructor function itself. I've explained this in more detail in my updates to the file.

Please correct me if I'm wrong.