Narrative: Explain the on() function before code splat
As was, the code sample was readable, but for
beginners, it was hard to figure out where they
ought to focus.

The sample (code splat) assumed that the beginners
could backwards-infer from the code what the
constraints ought have been and where JavaScript
would have been insufficiently robust in
documenting these constraints.

This occasions the removal of: